+++ src/parser.test.ts
@@ -1,6 +1,6 @@
 import html from "./testdata/amsterdamRent";
 import { htmlToJSDOM } from "./utils";
-import { parseListings } from "./scraper";
+import { parseListings } from "./parser";
 import {
   parseHousing,
   removeSpaces,
@@ -8,7 +8,7 @@ import {
   takeDetails,
   takePostalCode,
   takeRealEstate,
-} from ".";
+} from "./scraper";
 
 describe("Find house listings", () => {
   test("should find all ordered lists", () => {
