+++ src/parser.ts
@@ -1,0 +1,6 @@
+import { JSDOM } from "jsdom";
+
+export function parseListings(jsdom: JSDOM) {
+  const listedHouses = jsdom.window.document.querySelectorAll("ol");
+  return listedHouses;
+}
